% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Install RQDA Archive}
\usage{
install(verbose = FALSE)
}
\arguments{
\item{verbose}{logical.}
}
\description{
Install RQDA from CRAN while it is archived while performing an
installation of current and archived packages upon which it depends.
The key packages mentioned are RGtk2, gWidgets and gWidgetsRGtk2.
}
\note{
If Gtk+ is not properly installed, this function will stop
and require the user to to install it. Once this is done, this
function should be run again to complete the installation of the
packages that directly or indirectly depend on it.
}
