% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{make_FileList}
\alias{make_FileList}
\title{Create a "FileList" for an RQDA Project}
\usage{
make_FileList(files)
}
\arguments{
\item{files}{A character vector with the filepaths.}
}
\description{
In the RQDA package, there is a notion of a FileList, which is an R list in
which each element is a character vector of length 1 containing the content
of a text file and the \code{names} of the list are the filenames.
}
\details{
The \code{files} argument can be generated by \code{read_transcript}.
The approved format is plain text, but it may work for other text formats.
}
\note{
For this function to work properly, the target RQDA project has to be
open i.e. activated with \code{library(RQDA)} or \code{RQDA()} as the
case may be. Future version of this function may automate this process.
}
