% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_transcript}
\alias{read_transcript}
\title{Create Text Files from Word}
\usage{
read_transcript(destdir, docxfiles, ...)
}
\arguments{
\item{destdir}{Destination path. Where the files will be saved.}

\item{docxfiles}{Character vector containing the filepath(s) of
the Word documents to be converted.}

\item{...}{Arguments passed to \code{\link[readtext]{readtext}}.}
}
\value{
A character vector containing the paths of the generated text files.
}
\description{
Read text from a Word document and convert to plain text format
}
\details{
When \code{destdir} is unchanged, the function uses the current
working directory as the location for saving any files converted from
Word.
}
