% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\alias{install_rgtk2_and_deps}
\title{Install Archived RQDA}
\usage{
install(type = c("binary", "source"), verbose = FALSE)

install_rgtk2_and_deps(type = c("binary", "source"), verbose = FALSE)
}
\arguments{
\item{type}{A string, one of \code{binary} or \code{source}.}

\item{verbose}{logical.}
}
\description{
Install RQDA from CRAN archives and at the same time installing its core
dependencies mainly, RGtk2, cairoDevice, gWidgets, and gWidgetsRGtk2. Where
necessary, compatible versions of R and Rtools are also installed.
}
\details{
The \code{install} function carries out wholesale installation of
all the packages required by RQDA, before actually installing it.

\code{install_rgtk2_and_deps} installs RGtk2. Before doing so, it fetches
the Gtk+ distribution, making it available for RGtk2. When a source
installation is used, GTK is required for compiling RGtk2 (and cairoDevice);
for the binary installation, Gtk+ is required for proper usage of RGtk2.
Additionally, if compatible versions of R (i.e. 4.0 to 4.1) or Rtools
(4.0) are not available on the machine, these will also be installed, but
only in an interactive session (Windows only). This operations of this
function are a subset of \code{install} and thus when that function is
called, this one is called internally.
}
\note{
The binary option will download RGtk2 from non-CRAN repository, while
RQDA, cairoDevice, gWidgets, and gWidgetsRGtk2 are downloaded as source
packages from the CRAN archive.
}
\examples{
# install RQDA (and dependencies) using defaults
\dontrun{install()}

# install all from source, verbosely
\dontrun{install("source", TRUE)}

}
