% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_transcript}
\alias{read_transcript}
\title{Read text from a Wprd document and convert to plain text format}
\usage{
read_transcript(destdir, docxfiles)
}
\arguments{
\item{destdir}{Destination path. Where the files will be saved.}

\item{docxfiles}{Character vector containing the filepath(s) of
the Word documents to be converted.}
}
\value{
A character vector containing the paths of the generated text files.
}
\description{
Read text from a Wprd document and convert to plain text format
}
\details{
When \code{destdir} is unchanged, the function uses the current
working directory as the location for saving any files converted from
Word.
}
