% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_codes}
\alias{search_codes}
\title{Search for codes using a given term}
\usage{
search_codes(searchterm)
}
\arguments{
\item{searchterm}{A string.}
}
\value{
An object of class \code{data.frame}, which is subset of the coding
table.
}
\description{
Search for codes using a given term
}
